Fixed failing tests and added interface for `RequotingService`
