+++ src/test/kotlin/com/hedvig/underwriter/testhelp/QuoteClient.kt
@@ -2,7 +2,9 @@ package com.hedvig.underwriter.testhelp
 
 import com.hedvig.underwriter.model.Quote
 import com.hedvig.underwriter.web.dtos.CompleteQuoteResponseDto
+import com.hedvig.underwriter.web.dtos.QuoteBundleResponseDto
 import com.hedvig.underwriter.web.dtos.SignedQuoteResponseDto
+import com.hedvig.underwriter.web.dtos.SignedQuotesResponseDto
 import org.apache.commons.lang.RandomStringUtils
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.web.client.TestRestTemplate
@@ -121,7 +123,7 @@ class QuoteClient {
         birthdate: String = "1944-08-04",
         street: String = "ApStreet ${RandomStringUtils.randomNumeric(2)}",
         apartment: String = "1tv",
-        floor: Int = 1,
+        floor: String = "1",
         bbrid: String? = null,
         zip: String = "1234",
         city: String = "ApCity",
@@ -137,7 +139,7 @@ class QuoteClient {
         birthdate: String = "1944-08-04",
         street: String = "ApStreet ${RandomStringUtils.randomNumeric(2)}",
         apartment: String = "1tv",
-        floor: Int = 1,
+        floor: String = "1",
         bbrid: String? = null,
         zip: String = "1234",
         city: String = "ApCity",
@@ -152,47 +154,59 @@ class QuoteClient {
     fun createDanishAccidentQuote(
         birthdate: String = "1944-08-04",
         street: String = "ApStreet ${RandomStringUtils.randomNumeric(2)}",
+        apartment: String = "1tv",
+        floor: String = "1",
+        bbrid: String? = null,
         zip: String = "1234",
         city: String = "ApCity",
         coInsured: Int = 1,
         student: Boolean = false
     ): CompleteQuoteResponseDto {
-        return createDanishAccidentQuote<CompleteQuoteResponseDto>(birthdate, street, zip, city, coInsured, student).body!!
+        return createDanishAccidentQuote<CompleteQuoteResponseDto>(birthdate, street, apartment, floor, bbrid, zip, city, coInsured, student).body!!
     }
 
     fun createDanishAccidentQuoteRaw(
         birthdate: String = "1944-08-04",
         street: String = "ApStreet 11",
+        apartment: String = "1tv",
+        floor: String = "1",
+        bbrid: String? = null,
         zip: String = "1234",
         city: String = "ApCity",
         coInsured: Int = 1,
         student: Boolean = false
 
     ): ResponseEntity<String> {
-        return createDanishAccidentQuote<String>(birthdate, street, zip, city, coInsured, student)
+        return createDanishAccidentQuote<String>(birthdate, street, apartment, floor, bbrid, zip, city, coInsured, student)
     }
 
     fun createDanishTravelQuote(
         birthdate: String = "1944-08-04",
         street: String = "ApStreet ${RandomStringUtils.randomNumeric(2)}",
+        apartment: String = "1tv",
+        floor: String = "1",
+        bbrid: String? = null,
         zip: String = "1234",
         city: String = "ApCity",
         coInsured: Int = 1,
         student: Boolean = false
     ): CompleteQuoteResponseDto {
-        return createDanishTravelQuote<CompleteQuoteResponseDto>(birthdate, street, zip, city, coInsured, student).body!!
+        return createDanishTravelQuote<CompleteQuoteResponseDto>(birthdate, street, apartment, floor, bbrid, zip, city, coInsured, student).body!!
     }
 
     fun createDanishTravelQuoteRaw(
         birthdate: String = "1944-08-04",
         street: String = "ApStreet ${RandomStringUtils.randomNumeric(2)}",
+        apartment: String = "1tv",
+        floor: String = "1",
+        bbrid: String? = null,
         zip: String = "1234",
         city: String = "ApCity",
         coInsured: Int = 1,
         student: Boolean = false
 
     ): ResponseEntity<String> {
-        return createDanishTravelQuote<String>(birthdate, street, zip, city, coInsured, student)
+        return createDanishTravelQuote<String>(birthdate, street, apartment, floor, bbrid, zip, city, coInsured, student)
     }
 
     fun signQuote(
@@ -228,6 +242,28 @@ class QuoteClient {
 
     fun getQuote(quoteId: UUID): Quote? {
         return restTemplate.getForObject("/_/v1/quotes/$quoteId", Quote::class.java)
+    }
+
+    fun createBundle(vararg quoteIds: UUID): QuoteBundleResponseDto {
+
+        val ids = quoteIds.joinToString("\", \"", "\"", "\"")
+
+        val request = """
+            { "quoteIds": [$ids] }
+        """.trimIndent()
+
+        return postJson<QuoteBundleResponseDto>("/_/v1/quotes/bundle", request).body!!
+    }
+
+    fun createBundleRaw(vararg quoteIds: UUID): ResponseEntity<String> {
+
+        val ids = quoteIds.joinToString("\", \"", "\"", "\"")
+
+        val request = """
+            { "quoteIds": [$ids] }
+        """.trimIndent()
+
+        return postJson("/_/v1/quotes/bundle", request)
     }
 
     private inline fun <reified T : Any> createNorwegianHomeContentQuote(
@@ -378,7 +414,7 @@ class QuoteClient {
         birthdate: String,
         street: String,
         apartment: String,
-        floor: Int,
+        floor: String,
         bbrid: String? = null,
         zip: String,
         city: String,
@@ -420,6 +456,9 @@ class QuoteClient {
     private inline fun <reified T : Any> createDanishAccidentQuote(
         birthdate: String,
         street: String,
+        apartment: String,
+        floor: String,
+        bbrid: String? = null,
         zip: String,
         city: String,
         coInsured: Int,
@@ -437,6 +476,9 @@ class QuoteClient {
                 "incompleteQuoteData": {
                     "type": "danishAccident",
                     "street": "$street",
+                    "apartment": "$apartment",
+                    "floor": "$floor",
+                    "bbrid": "$bbrid",
                     "zipCode": "$zip",
                     "city": "$city",
                     "coInsured": $coInsured,
@@ -453,6 +495,9 @@ class QuoteClient {
     private inline fun <reified T : Any> createDanishTravelQuote(
         birthdate: String,
         street: String,
+        apartment: String,
+        floor: String,
+        bbrid: String? = null,
         zip: String,
         city: String,
         coInsured: Int,
@@ -470,6 +515,9 @@ class QuoteClient {
                 "incompleteQuoteData": {
                     "type": "danishTravel",
                     "street": "$street",
+                    "apartment": "$apartment",
+                    "floor": "$floor",
+                    "bbrid": "$bbrid",
                     "zipCode": "$zip",
                     "city": "$city",
                     "coInsured": $coInsured,
@@ -507,6 +555,70 @@ class QuoteClient {
         """.trimIndent()
 
         return postJson("/_/v1/quotes/$quoteId/sign", request)
+    }
+
+    fun signQuoteBundle(
+        firstName: String = "Apan",
+        lastName: String = "Apansson",
+        ssn: String?,
+        email: String = "apan@apansson.se",
+        startDate: LocalDate = LocalDate.now(),
+        price: Double?,
+        currency: String?,
+        vararg quoteIds: UUID
+    ): SignedQuotesResponseDto {
+
+        val currencyString = currency?.let { "\"" + currency + "\"" }
+        val ssnString = ssn?.let { "\"" + ssn + "\"" }
+
+        val request = """
+            {
+                "name": {
+                    "firstName": "$firstName",
+                    "lastName": "$lastName"
+                },
+                "ssn": $ssnString,
+                "startDate": "$startDate",
+                "email": "$email",
+                "quoteIds": [${quoteIds.joinToString("\", \"", "\"", "\"")}],
+                "price": $price,
+                "currency": $currencyString
+            }
+        """.trimIndent()
+
+        return postJson<SignedQuotesResponseDto>("/_/v1/quotes/bundle/signFromRapio", request).body!!
+    }
+
+    fun signQuoteBundleRaw(
+        firstName: String = "Apan",
+        lastName: String = "Apansson",
+        ssn: String?,
+        email: String = "apan@apansson.se",
+        startDate: LocalDate = LocalDate.now(),
+        price: Double?,
+        currency: String?,
+        vararg quoteIds: UUID
+    ): ResponseEntity<String> {
+
+        val currencyString = currency?.let { "\"" + currency + "\"" }
+        val ssnString = ssn?.let { "\"" + ssn + "\"" }
+
+        val request = """
+            {
+                "name": {
+                    "firstName": "$firstName",
+                    "lastName": "$lastName"
+                },
+                "ssn": $ssnString,
+                "startDate": "$startDate",
+                "email": "$email",
+                "quoteIds": [${quoteIds.joinToString("\", \"", "\"", "\"")}],
+                "price": $price,
+                "currency": $currencyString
+            }
+        """.trimIndent()
+
+        return postJson("/_/v1/quotes/bundle/signFromRapio", request)
     }
 
     private inline fun <reified T : Any> postJson(url: String, data: String): ResponseEntity<T> {
